Typo (I think) in the
In the file eeckhoutKircher2013.py, I tried to make a workers class and
examine the properties of the scaled pdf and cdf. I think the scaling
factor definition is wrong, because let's say, I have 1000 workers. Then
self.norm_const delivers F(xmax)-F(xmin). This probably is close to 1.
Then self.scaling_factor divides the self.size / self.norm_const, which
in this case would be 1000/1. And then when calculating the scaled pdf,
it multiplies F(x) by this scaling factor (which is 1000)!
In the notebook it is not an issue, because you set the measure to 1, so
the scaling factor is close to 1 anyways.
In the new inputs.py file we merged yesterday this issue is corrected,
so probably this is a typo of a file that you may not be even using
anymore. But in any case, it is fixed now.