Added support for different penalties.
New two optional arguments for Calculating_MSE and
SttubornObjectiveFunction: penalty_wp (penalty for obsevations out fo
range in profits and wages) and penalty_cdf (penalty for observations
out of range in the cdf for sizes. Soon I will also update the notebook
to show examples of how to call it.